Add files via upload
Resuts:
Obtained values of precision, recall, F1 score and accuracy for classification of ML  models. Naive Bayes models are the least suited for the image classification of brain tumors. It can be seen that Random forest shows best results due to its improved generalization and  embedded ensemble learning feature.
Obtained values of precision, recall, F1 score and accuracy for classification of four DL models. The performance is highest for our proposed CNN model with an accuracy of 92% which is more than all other models trained

Application:
Magnetic resonance (MR) imaging and computed  tomography (CT) scans of the brain are the two most general tests to check the existence of a tumor as of now. Radiologists have to manually check and test the positioning and type of tumor based on the MRIs. This process of classifying brain tumors is both tedious and time consuming. So in order to  decrease the level of complexities involved in brain tumors and their properties, this model has been proposed. Application of automated classiﬁcation techniques using Machine Learning and Artiﬁcial Intelligence has consistently shown higher accuracy than manual classiﬁcation. Thus we propose our CNN model to classify images in order to help medical specialists in automated classification of tumors.